pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orbit"
version = "0.1.0"
description = "Apple iCloud ↔ Skylight calendar sync service"
authors = [
    {name = "Orbit Team"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "caldav>=1.3.0",
    "python-dateutil>=2.8.0",
    "pydantic[dotenv]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "structlog>=23.2.0",
    "apscheduler>=3.10.0",
    "prometheus-client>=0.19.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pytz>=2023.3",
    "requests>=2.31.0",
    "bcrypt>=3.2.0",
    "ulid-py>=1.1.0",
    "jsonschema>=4.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.26.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "fastmcp>=2.12.0",
    "PyYAML>=6.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [
    "deprecated/",
    "scripts/",
    "ui/node_modules/",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
